+++ src/expect/mod.rs
@@ -1,0 +1,5 @@
+mod expect_none;
+mod unwrap_never;
+
+pub use expect_none::ExpectNone;
+pub use unwrap_never::UnwrapNever;