+++ src/internal.rs
@@ -1,7 +1,0 @@
-pub trait Sealed {}
-
-impl<T, E> Sealed for Result<T, E> {}
-
-impl<T> Sealed for Option<T> {}
-
-impl Sealed for bool {}