+++ src/sink/sink_option.rs
@@ -1,13 +1,12 @@
-use crate::internal::Sealed;
-
 /// An extension for handling [`Some`] variants by sinking them into a side-effecting function.
-pub trait SinkOption<T>: Sealed {
+#[sealed::sealed]
+pub trait SinkOption<T> {
     /// Handles [`Some`] variants of [`Option`]s by sinking them into `sink`.
     ///
     /// # Example
     ///
     /// ```rust
-    /// use fluent_result::SinkOption;
+    /// use fluent_result::sink::SinkOption;
     ///
     /// let mut log = String::new();
     /// Some("oops").sink(|e| log.push_str(e));
@@ -18,6 +17,7 @@ pub trait SinkOption<T>: Sealed {
         F: FnOnce(T);
 }
 
+#[sealed::sealed]
 impl<T> SinkOption<T> for Option<T> {
     fn sink<F>(self, sink: F)
     where