+++ tests/unwrap-never.rs
@@ -1,5 +1,6 @@
-use fluent_result::UnwrapNever;
 use std::convert::Infallible;
+
+use fluent_result::expect::UnwrapNever;
 
 #[test]
 fn unwrap_never() {